requirements.txt
fastmcp>=0.3.3
playwright>=1.44
openai>=1.0.0
readability>=0.3.1
beautifulsoup4>=4.12.0
lxml>=5.0.0
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
